+++ src/components/molecules/Board/BoardFooter/index.tsx
