+++ src/pages/boards/[board_type]/new/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import { SubmitHandler, useForm } from 'react-hook-form';
 import CommonCenterWrapper from '@src/components/common/Layout/CommonCenterWrapper';
-import QuillEditor from '@src/components/pages/boardsPage/Editor/QuillEditor';
+import QuillEditor from '@src/components/molecules/Editor/QuillEditor';
 import { useRouter } from 'next/router';
 import { KeyboardEvent } from 'react';
 import axios from 'axios';
