+++ src/pages/login/register/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/jsx-props-no-spreading */
 
-import RegisterForm from '@src/components/pages/loginPage/RegisterForm';
+import RegisterForm from '@src/components/pages/login/RegisterForm';
 
 export default function Register() {
   return <RegisterForm />;
